new Tagging interface TorrentScrapedFileSearchResult.
ExtratorrentCrawledSearchResult and KATCrawledSearchResult
implement this interface, and are now named

ExtratorrentScrapedFileSearchResult and KATScrapedFileSearchResult.

Initially I thought this would be simply a tagging interface to avoid
those ugly instanceof hacks per implementation everywhere, in the process
I noticed those 2 classes had a getFilePath() method that wasn't in
any interface, I've put it here for now.

I suppose getFilePath() has to do with the internal path of the file
inside the torrent info. Current implementations are only returning
the file name, as this is all that can be easily gathered from a
page scrape.

Perhaps, to make the integration easier with clients when they're about to start
the related partial download, there should be a way to either
update the ScrapedFileSearchResult instance, or to immediatly re-create
a new instance with the actual path of the file once we have the torrent info.